lab_4/rent/requirements.txt
celery==3.1.25
Django==2.2.17
django-redis==4.12.1
django-rest-framework==0.1.0
djangorestframework==3.12.2

stripe
win10toast
